+++ src/train.py
@@ -259,7 +259,7 @@ def main():
 
         # Get prediction nodes
         placeholder_z = tf.placeholder(tf.float32, shape=sample_z.shape)
-        placeholder_y = tf.placeholder(tf.int32, shape=sample_y.shape)    #Remember
+        placeholder_y = tf.placeholder(tf.float32, shape=sample_y.shape)    #Remember
         
         if params.get('is_accompaniment'):
             c_shape = np.append(sample_x.shape[:-1], 1)
